fix: clean up and declare types for react-docgen
